pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "factoriocalc"
version = "0.0.1"
authors = [
  { name="Kevin Atkinson" },
]
description = "A python module to help you plan your factory for Factorio."
keywords = ["Factorio"]
readme = "README.rst"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
]
dependencies = [
  "factorio-draftsman"
]


[project.urls]
"Homepage" = "https://github.com/kevina/factoriocalc"
"Bug Tracker" = "https://github.com/kevina/factoriocalc/issues"
